+++ frontend/src/lib/components/InputTransformForm.svelte
@@ -407,6 +407,8 @@
 					bind:pickForField
 					showSchemaExplorer
 					nullable={schema.properties[argName].nullable}
+					bind:title={schema.properties[argName].title}
+					bind:placeholder={schema.properties[argName].placeholder}
 				/>
 			{:else if arg.expr != undefined}
 				<div class="border mt-2">