+++ frontend/src/lib/components/LightweightArgInput.svelte
@@ -51,6 +51,8 @@
 	export let hideResourceInput: boolean = false
 	export let resourceInputUnsupported: boolean = false
 	export let render = true
+	export let title: string | undefined = undefined
+	export let placeholder: string | undefined = undefined
 
 	const dispatch = createEventDispatcher()
 
@@ -180,8 +182,8 @@
 			<div>
 				{#if displayHeader}
 					<FieldHeader
-						prettify
-						{label}
+						prettify={emptyString(title)}
+						label={title && !emptyString(title) ? title : label}
 						{required}
 						{type}
 						{contentEncoding}
@@ -231,7 +233,7 @@
 										? ''
 										: 'border !border-red-700 !border-opacity-70 focus:!border-red-700 focus:!border-opacity-30'
 								)}
-								placeholder={defaultValue ?? ''}
+								placeholder={placeholder ?? defaultValue ?? ''}
 								bind:value
 								min={extra['min']}
 								max={extra['max']}
@@ -430,7 +432,7 @@
 							class={valid
 								? ''
 								: 'border border-red-700 border-opacity-30 focus:border-red-700 focus:border-opacity-3'}
-							placeholder={defaultValue ?? ''}
+							placeholder={placeholder ?? defaultValue ?? ''}
 							bind:value
 						/>
 					{:else if inputCat == 'currency'}
@@ -439,7 +441,7 @@
 							class={valid
 								? ''
 								: 'border border-red-700 border-opacity-30 focus:border-red-700 focus:border-opacity-3'}
-							placeholder={defaultValue ?? ''}
+							placeholder={placeholder ?? defaultValue ?? ''}
 							bind:value
 						/>
 					{:else if inputCat == 'string'}
@@ -458,7 +460,7 @@
 										class="col-span-10 {valid && error == ''
 											? ''
 											: 'border !border-red-700 !border-opacity-70 focus:!border-red-700 focus:!border-opacity-30'}"
-										placeholder={defaultValue ?? ''}
+										placeholder={placeholder ?? defaultValue ?? ''}
 										bind:value
 										on:pointerdown|stopPropagation={(e) => {
 											dispatch('inputClicked', e)